Replace computeDistanceIndex with generic getDistanceTo
This accepts a condition which it looks for in a sector. This allows
the distanceIndex calculation for buying and selling to be easily
separated, simply by changing the function that checks for a valid
sector, and could also be used for calculating the distance between
two sectors by having the condition check `num`. This will be useful
for calculating the credits per turn value when generating routes.