Handle the increased turn consumption of warps
This is done by doing a sorted insertion on toVisit, keeping it
ordered by distance at all times. This way, new warps will be pushed
to the end of the list, and any adjacent sectors that would be faster
than warping are inserted ahead of it. This way, the search should
always explore the closest sectors first and only consider warps when
all closer options have been exhausted.